Added Python 3 support

+++ plugins/dstat_disk_avgqu.py
@@ -25,7 +25,7 @@ class dstat_plugin(dstat):
             ret.append(name)
         for item in objlist: ret.append(item)
         if not ret:
-            raise Exception, "No suitable block devices found to monitor"
+            raise Exception('No suitable block devices found to monitor')
         return ret
 
     def vars(self):
@@ -56,7 +56,7 @@ class dstat_plugin(dstat):
             name = l[2]
             if name not in self.vars or name == 'total': continue
             self.set2[name] = dict(
-                rq_ticks = long(l[13]),
+                rq_ticks = int(l[13]),
             )
 
         for name in self.vars:
