+++ plugins/dstat_disk_avgrq.py
@@ -26,7 +26,7 @@ class dstat_plugin(dstat):
             ret.append(name)
         for item in objlist: ret.append(item)
         if not ret:
-            raise Exception, "No suitable block devices found to monitor"
+            raise Exception('No suitable block devices found to monitor')
         return ret
 
     def vars(self):
@@ -57,9 +57,9 @@ class dstat_plugin(dstat):
             name = l[2]
             if name not in self.vars or name == 'total': continue
             self.set2[name] = dict(
-                nr_ios = long(l[3])+long(l[7]),
-                rd_sect = long(l[9]),
-                wr_sect = long(l[11]),
+                nr_ios = int(l[3])+int(l[7]),
+                rd_sect = int(l[9]),
+                wr_sect = int(l[11]),
             )
 
         for name in self.vars:
