+++ plugins/dstat_ib.py
@@ -46,7 +46,7 @@ class dstat_plugin(dstat):
             if name in self.discover + ['total']:
                 ret.append(name)
         if not ret:
-            raise Exception, "No suitable network interfaces found to monitor"
+            raise Exception('No suitable network interfaces found to monitor')
         return ret
 
     def name(self):
@@ -66,8 +66,8 @@ class dstat_plugin(dstat):
             xmit_lines = dopen(xmit_counter_name).readlines()
             if len(rcv_lines) < 1 or len(xmit_lines) < 1:
                 continue
-            rcv_value = long(rcv_lines[0])
-            xmit_value = long(xmit_lines[0])
+            rcv_value = int(rcv_lines[0])
+            xmit_value = int(xmit_lines[0])
             if name in self.vars :
                 self.set2[name] = (rcv_value, xmit_value)
             self.set2['total'] = ( self.set2['total'][0] + rcv_value, self.set2['total'][1] + xmit_value)
