+++ plugins/dstat_net_packets.py
@@ -40,7 +40,7 @@ class dstat_plugin(dstat):
             if name in self.discover + ['total', 'lo']:
                 ret.append(name)
         if not ret:
-            raise Exception, "No suitable network interfaces found to monitor"
+            raise Exception('No suitable network interfaces found to monitor')
         return ret
 
     def name(self):
@@ -53,9 +53,9 @@ class dstat_plugin(dstat):
             if l[2] == '0' and l[10] == '0': continue
             name = l[0]
             if name in self.vars :
-                self.set2[name] = ( long(l[2]), long(l[10]) )
+                self.set2[name] = ( int(l[2]), int(l[10]) )
             if not self.totalfilter.match(name):
-                self.set2['total'] = ( self.set2['total'][0] + long(l[2]), self.set2['total'][1] + long(l[10]))
+                self.set2['total'] = ( self.set2['total'][0] + int(l[2]), self.set2['total'][1] + int(l[10]))
 
         if update:
             for name in self.set2:
