+++ plugins/dstat_rpc.py
@@ -14,7 +14,7 @@ class dstat_plugin(dstat):
         for l in self.splitlines():
             if not l or l[0] != 'rpc': continue
             for i, name in enumerate(self.vars):
-                self.set2[name] = long(l[i+1])
+                self.set2[name] = int(l[i+1])
 
         for name in self.vars:
             self.val[name] = (self.set2[name] - self.set1[name]) * 1.0 / elapsed
