+++ plugins/dstat_zfs_zil.py
@@ -19,7 +19,7 @@ class dstat_plugin(dstat):
             l[0].split()
             name = l[0]
             if name in self.vars:
-                self.set2[name] = long(l[2])
+                self.set2[name] = int(l[2])
 
         for i, name in enumerate (self.vars):
             if self.counter[i]:
